backend/requirements.txt
fastapi>=0.104.0,<0.105.0
uvicorn>=0.23.2,<0.24.0
python-multipart>=0.0.6,<0.1.0
pandas>=2.3.1,<3.0.0
openpyxl>=3.1.5,<4.0.0
python-dotenv>=1.1.1,<2.0.0
openai>=1.3.0,<2.0.0
websockets>=12.0,<13.0
